# public/api-docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Cipi API",
...
        "version": "1.17.0",
        "contact": {
            "name": "Cipi",
            "url": "https://cipi.sh"
        },
        "license": {
            "name": "MIT"
        }
    },
    "servers": [
        {
            "url": "/api",
            "description": "API"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/apps/{name}": {
            "get": {
                "tags": [
                    "Apps"
                ],
                "summary": "Show app",
                "description": "Show app details. Returns data synchronously. Requires `apps-view` ability.",
                "operationId": "appShow",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "App details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AppShowResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Apps"
                ],
                "summary": "Edit app",
                "description": "Edit an app (PHP version, branch, repository, or primary domain). Validates app exists synchronously; validates domain format and uniqueness (excluding the current app's aliases) before dispatch. Changing the primary domain moves the old one to aliases, updates Nginx, `APP_URL`, webhooks, and re-issues SSL when applicable — requires [Cipi 4.6.2+](https://github.com/cipi-sh/cipi/releases/tag/4.6.2). Dispatches CLI command as async job. Requires `apps-edit` ability.",
                "operationId": "appEdit",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AppEditRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Job dispatched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobAcceptedResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/Conflict"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Apps"
                ],
                "summary": "Delete app",
                "description": "Permanently delete an app and all associated resources. Validates app exists synchronously. Dispatches CLI command as async job. Requires `apps-delete` ability.",
                "operationId": "appDelete",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AppName"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Job dispatched",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobAcceptedResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Sanctum API token. Create with: cipi api token create"
            }
        },
        "parameters": {
            "AppName": {
                "name": "name",
                "in": "path",
                "required": true,
                "description": "App username (e.g. myapp)",
                "schema": {
                    "type": "string",
                    "pattern": "^[a-z][a-z0-9]{2,31}$"
                }
            },
...
        },
        "schemas": {
...
            "AppEditRequest": {
                "type": "object",
                "properties": {
                    "php": {
                        "type": "string",
                        "enum": [
                            "7.4",
                            "8.0",
                            "8.1",
                            "8.2",
                            "8.3",
                            "8.4",
                            "8.5"
                        ],
                        "description": "PHP version"
                    },
                    "branch": {
                        "type": "string",
                        "description": "Git branch",
                        "maxLength": 64
                    },
                    "repository": {
                        "type": "string",
                        "description": "Git repository URL"
                    },
                    "domain": {
                        "type": "string",
                        "description": "New primary domain. The previous primary becomes an alias; promoting an existing alias to primary is supported. Requires Cipi 4.6.2+.",
                        "example": "new.example.com"
                    }
                }
            },
...
            "AppShowResponse": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "object",
                        "properties": {
                            "app": {
                                "type": "string"
                            },
                            "domain": {
                                "type": "string"
                            },
                            "php": {
                                "type": "string"
                            },
                            "branch": {
                                "type": "string"
                            },
                            "repository": {
                                "type": "string"
                            },
                            "aliases": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "suspended": {
                                "type": "boolean"
                            },
                            "basic_auth": {
                                "type": "boolean",
                                "description": "Whether HTTP Basic Auth is enabled on the app vhost"
                            },
                            "created_at": {
                                "type": "string"
                            },
                            "engine": {
                                "type": "string",
                                "nullable": true,
                                "enum": [
                                    "mariadb",
                                    "pgsql"
                                ],
                                "description": "App database engine (Laravel apps; null for custom / unset)",
                                "example": "mariadb"
                            },
                            "www_redirect": {
                                "type": "string",
                                "nullable": true,
                                "enum": [
                                    "to-root",
                                    "from-root"
                                ],
                                "description": "Canonical www redirect mode (null when none). Cipi 4.8+.",
                                "example": "to-root"
                            },
                            "force_https": {
                                "type": "boolean",
                                "description": "Whether HTTP → HTTPS redirect is forced (set by ssl install/force). Cipi 4.8+.",
                                "example": true
                            },
                            "octane": {
                                "type": "string",
                                "nullable": true,
                                "enum": [
                                    "frankenphp"
                                ],
                                "description": "Octane server when the app uses Octane instead of PHP-FPM (null for FPM apps). Cipi 5.0+.",
                                "example": "frankenphp"
                            },
                            "octane_port": {
                                "type": "integer",
                                "nullable": true,
                                "description": "Localhost port for the Octane process (null for FPM apps). Cipi 5.0+.",
                                "example": 8100
                            }
                        }
                    }
                }
            },
...
            "JobAcceptedResponse": {
                "type": "object",
                "properties": {
                    "job_id": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Background job UUID. Poll GET /api/jobs/{id} for status.",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending"
                        ],
                        "example": "pending"
                    }
                }
            },
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "description": "Error message"
                    }
                }
            },
...
        },
        "responses": {
            "Unauthorized": {
                "description": "Missing or invalid token",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated."
                                }
                            }
                        }
                    }
                }
            },
            "Forbidden": {
                "description": "Token lacks required ability",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Invalid ability provided."
                                }
                            }
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "Conflict": {
                "description": "Resource conflict (e.g. domain or app already exists, or duplicate pending database create)",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ErrorResponse"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "errors": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
...
        }
    },
...
}